+++ packages/validator/src/rules/all.ts
@@ -1,7 +1,6 @@
-import { allP } from '../helpers/promise.helper';
-import { validatorSuccess, validatorError } from '../index';
-
-export type BoolExpCallback = (property: any) => boolean | Promise<boolean>;
+import { allP } from '../helpers';
+import { validatorSuccess, validatorError } from '../helpers';
+import { BoolExpCallback } from '../types';
 
 export const all = (evaluator: BoolExpCallback) => async (value: any[]) =>
   (await allP(async (elem: any) => await evaluator(elem), value))
