+++ packages/validator/src/rules/array.ts
@@ -1,0 +1,4 @@
+import { wrap } from '../helpers';
+import { type } from './type';
+
+export const array = wrap(type('Array'));
