+++ packages/validator/src/rules/boolean.ts
@@ -1,0 +1,4 @@
+import { wrap } from '../helpers';
+import { type } from './type';
+
+export const boolean = wrap(type('boolean'));
