+++ packages/validator/src/rules/exists.ts
@@ -1,4 +1,5 @@
-import { validatorSuccess, validatorError, ValidatorReturn } from '../index';
+import { validatorSuccess, validatorError } from '../helpers';
+import { ValidatorReturn } from '../types';
 
 export function exists(existingValues: any[]) {
   return (value: any): ValidatorReturn =>
