+++ packages/validator/src/rules/hexColor.ts
@@ -1,0 +1,4 @@
+import { wrap } from '../helpers';
+import { regex } from './regex';
+
+export const webColor = wrap(regex('^#[0-9a-fA-F]{6}$'));
