+++ packages/validator/src/rules/length.ts
@@ -1,5 +1,4 @@
-import * as R from 'ramda';
-import { validatorSuccess, validatorError } from '../index';
+import { validatorSuccess, validatorError } from '../helpers';
 
 export const length = (
   min: number,
@@ -7,7 +6,7 @@ export const length = (
   errorMessage: string | undefined = undefined,
 ) => (value: any[] | string) => {
   if (typeof value === 'string') {
-    value = R.trim(value);
+    value = value.trim();
   }
   return value.length <= max && value.length >= min
     ? validatorSuccess()
@@ -15,10 +14,5 @@ export const length = (
         errorMessage || `needs to have a length between ${min} and ${max}`,
       );
 };
-
-export const min = (min: number) =>
-  length(min, Infinity, `needs to have length of at least ${min}`);
-export const max = (max: number) =>
-  length(0, max, `cannot exceed length of ${max}`);
 
 export default length;
