+++ packages/validator/src/rules/max.ts
@@ -1,0 +1,4 @@
+import { length } from './length';
+
+export const max = (max: number) =>
+  length(0, max, `cannot exceed length of ${max}`);
