+++ packages/validator/src/rules/min.ts
@@ -1,0 +1,4 @@
+import { length } from './length';
+
+export const min = (min: number) =>
+  length(min, Infinity, `needs to have length of at least ${min}`);
