+++ packages/validator/src/rules/number.ts
@@ -1,0 +1,4 @@
+import { wrap } from '../helpers';
+import { type } from './type';
+
+export const number = wrap(type('number'));
