+++ packages/validator/src/rules/object.ts
@@ -1,23 +1,15 @@
-import validate, {
-  ValidatorReturn,
-  validatorSuccess,
-  validatorError,
-} from '../index';
-
-export interface RuleObject {
-  [key: string]: ((...args: any[]) => ValidatorReturn)[];
-}
+import { validatorSuccess, validatorError } from '../helpers';
+import { ValidatorReturn, RuleObject } from '../types';
+import validate from '../index';
 
 async function applyRules(rules: RuleObject, obj: object) {
   const { errors } = await validate(obj, rules);
   return errors;
 }
 
-const object = (rules: RuleObject = {}) => {
+export const object = (rules: RuleObject = {}) => {
   return async (obj: object): ValidatorReturn => {
     const formErrors = await applyRules(rules, obj);
     return !formErrors ? validatorSuccess() : validatorError(formErrors);
   };
 };
-
-export default object;
