+++ packages/validator/src/rules/string.ts
@@ -1,0 +1,4 @@
+import { wrap } from '../helpers';
+import { type } from './type';
+
+export const string = wrap(type('string'));
