+++ packages/validator/src/rules/uuid4.ts
@@ -1,0 +1,8 @@
+import { regex } from './regex';
+import { wrap } from '../helpers';
+
+export const uuid = wrap(
+  regex(
+    '[0-9a-f]{8}-?[0-9a-f]{4}-?[1-5][0-9a-f]{3}-?[89ab][0-9a-f]{3}-?[0-9a-f]{12}',
+  ),
+);
